pyproject.toml
[tool.poetry]
name = "django-ulogin"
version = "1.1.1"
description = "User social authentication with ulogin.ru service"
readme = "README.rst"
authors = ["Mikhail Porokhovnichenko <marazmiki@gmail.com>"]
license = "MIT"
classifiers = [
        "Development Status :: 5 - Production/Stable",
        "Environment :: Web Environment",
        "Framework :: Django",
        "Framework :: Django :: 2.2",
        "Framework :: Django :: 3.0",
        "Framework :: Django :: 3.1",
        "Framework :: Django :: 3.2",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: MIT License",
        "Operating System :: OS Independent",
        "Programming Language :: Python",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.5",
        "Programming Language :: Python :: 3.6",
        "Programming Language :: Python :: 3.7",
        "Programming Language :: Python :: 3.8",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
        "Topic :: Internet :: WWW/HTTP :: Session"
]


[tool.poetry.dependencies]
python = "^3.6 || ^3.7 || ^3.8 || ^3.9 || ^3.10"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
bump2version = "^1.0.0"
flake8 = "^3.8.3"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-isort = "^2.0.0"
pytest-flake8 = "^1.0.7"
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
